Model/classifier/requirements.txt
google-generativeai==0.5.4
python-dotenv==1.0.1